third_party/WebKit/LayoutTests/http/tests/inspector/network/network-xhr-async-response-type-blob.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="../network-test.js"></script>
<script>
function test()
{
    InspectorTest.recordNetwork();
    InspectorTest.makeXHR("GET", "resources/resource.php", true, undefined, undefined, [], false, undefined, 'blob', step2);

    function step2()
    {
        var request1 = InspectorTest.networkRequests().pop();
        InspectorTest.addResult(request1.url());
        InspectorTest.addResult("resource.type: " + request1.resourceType());
        InspectorTest.addResult("resource.size: " + request1.resourceSize);
        InspectorTest.addResult("resource.content before requesting content: " + request1.content);
        InspectorTest.assertTrue(!request1.failed, "Resource loading failed.");
        request1.requestContent().then(step3);
    }

    function step3()
    {
        var request1 = InspectorTest.networkRequests().pop();
        InspectorTest.addResult("resource.content after requesting content: " + request1.content);

        InspectorTest.completeTest();
    }
}
</script>
</head>
<body onload="runTest()">
 <p>
 Tests XHR network resource type and size for asynchronous requests when "blob" is specified as the response type.
</p>
</body>
</html>
