third_party/WebKit/LayoutTests/http/tests/inspector/network/network-xhr-async-response-type-blob-expected.txt

CONSOLE MESSAGE: line 6: XHR loaded: 1
Tests XHR network resource type and size for asynchronous requests when "blob" is specified as the response type.

http://127.0.0.1:8000/inspector/network/resources/resource.php
resource.type: xhr
resource.size: 11
resource.content after requesting content: Hello world
